Turn log models into proper Json objects. Finish refactoring.
