+++ RatTracker-WPF/RatTracker-WPF/Models/CmdrJournal/HullDamageLog.cs
@@ -1,22 +1,27 @@
 ﻿using System;
+using Newtonsoft.Json;
 
-namespace RatTracker_WPF.Models.CmdrLog
+namespace RatTracker_WPF.Models.CmdrJournal
 {
+    [JsonObject]
     public class HullDamageLog : ICmdrJournalEntry
     {
         /// <summary>
         ///     Current health of the player;
         /// </summary>
+        [JsonProperty]
         public double Health { get; set; }
 
         /// <summary>
         ///     Time the event occured.
         /// </summary>
+        [JsonProperty("timestamp")]
         public DateTime Timestamp { get; set; }
 
         /// <summary>
         ///     Event type name.
         /// </summary>
+        [JsonProperty("event")]
         public string Event { get; set; }
     }
 }
