+++ RatTracker-WPF/RatTracker-WPF/Models/CmdrJournal/WingAddLog.cs
@@ -1,22 +1,27 @@
 ﻿using System;
+using Newtonsoft.Json;
 
-namespace RatTracker_WPF.Models.CmdrLog
+namespace RatTracker_WPF.Models.CmdrJournal
 {
+    [JsonObject]
     public class WingAddLog : ICmdrJournalEntry
     {
         /// <summary>
         ///     Name of the CMDR added to the wing.
         /// </summary>
+        [JsonProperty]
         public string Name { get; set; }
 
         /// <summary>
-        ///     Time the event occured
+        ///     Time the event occured.
         /// </summary>
+        [JsonProperty("timestamp")]
         public DateTime Timestamp { get; set; }
 
         /// <summary>
-        ///     Name of the event as seen in the cmdr log.
+        ///     Event type name.
         /// </summary>
+        [JsonProperty("event")]
         public string Event { get; set; }
     }
 }
