+++ RatTracker-WPF/RatTracker-WPF/Models/CmdrJournal/WingJoinLog.cs
@@ -1,27 +1,27 @@
 ﻿using System;
+using Newtonsoft.Json;
 
-namespace RatTracker_WPF.Models.CmdrLog
+namespace RatTracker_WPF.Models.CmdrJournal
 {
-    public class EscapeInterdictionLog : ICmdrJournalEntry
+    [JsonObject]
+    public class WingJoinLog : ICmdrJournalEntry
     {
         /// <summary>
-        ///     Name of the interdictor
+        ///     Array of Cmdr names who are also in the wing.
         /// </summary>
-        public string Interdictor { get; set; }
-
-        /// <summary>
-        ///     Whether the interdictor is a player.
-        /// </summary>
-        public bool IsPlayer { get; set; }
+        [JsonProperty]
+        public string[] Others { get; set; }
 
         /// <summary>
         ///     Time the event occured.
         /// </summary>
+        [JsonProperty("timestamp")]
         public DateTime Timestamp { get; set; }
 
         /// <summary>
         ///     Event type name.
         /// </summary>
+        [JsonProperty("event")]
         public string Event { get; set; }
     }
 }
