--- RatTracker-WPF/RatTracker-WPF/Models/CmdrLog/WingJoinLog.cs
@@ -1,22 +1,0 @@
-﻿using System;
-
-namespace RatTracker_WPF.Models.CmdrLog
-{
-    public class WingJoinLog : ICmdrJournalEntry
-    {
-        /// <summary>
-        ///     Array of Cmdr names who are also in the wing.
-        /// </summary>
-        public string[] Others { get; set; }
-
-        /// <summary>
-        ///     Time the event occured
-        /// </summary>
-        public DateTime Timestamp { get; set; }
-
-        /// <summary>
-        ///     Name of the event as seen in the cmdr log.
-        /// </summary>
-        public string Event { get; set; }
-    }
-}
